Nicely align tag numbers
The output of the show_entries and show_entries_all_lists functions is now fetched by the align_output function. This function finds the largest tag width, and uses that information to nicely align all tags. The changes may appear to be a bit cumbersome, but I think this is the only way to do it, as we need to generate all lines first, to know which line holds the largest tag.