Added fixed width formatting for the tag column
This addition was motivated by edofrederix's nice patch to dynamically evaluate the required maximum column width and format the column accordingly. I optedfor a fix width format to keep things simple with fewer operations. The dynamic approach may prove to be the way to go in the future if a fixed width becomes insufficient.