Added email and configuration file support.
The email support is similar to the print support except it mails the
list information to MAIL_RECIPIENT (defaults to
${USER}@localhost). The system must have a proper mailer setup for
external routing (SMTP).

gsd now looks for a configuration file (~/.gsd/config) which currently
holds DARK_BACKGROUND and MAIL_RECIPIENT. The value for MAIL_RECIPIENT
is a comma delimited list.