Added the archive command
The archive command moves all completed entries from a list and places
into the ARCHIVE list. The ARCHIVE list is a special list and is not
included in any list operations (e.g., "gsd sl"). To interact with the
archive list, you'll need to provide the '-l' option to gsd. For
example to see the items in the ARCHIVE list use:

  gsd -l ARCHIVE show

To remove all enties from the ARCHIVE, you use:

  gsd -l ARCHIVE purge